requirements.txt

flask
streamlit
pdfplumber
scikit-learn
openai>=1.0.0
python-dotenv
